StatisticsService/requirements.txt
clickhouse-driver
kafka-python==2.0.2
flask_sqlalchemy==3.1.1
Werkzeug==3.0.1
psycopg2==2.9.9
grpcio==1.51.1
grpcio-tools==1.51.1